backend/requirements.txt
# VeriMark AI - Python Dependencies

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
torch==2.9.1
torchvision==0.24.1
numpy==1.26.4
scikit-learn>=1.4.2

# Image Processing
pillow

# Security & Encryption
cryptography==42.0.5

# Data Processing
joblib==1.4.2

# Utilities
python-dotenv>=1.0.0
pydantic==2.7.4

# Optional: For production deployment
# gunicorn==21.2.0